requirements.txt
fastapi
uvicorn[standard]
jinja2
python-pam
python-multipart
aiofiles
jinja2
pydantic
